refactor: 도메인 클래스와 이름이 맞지않는 클래스 이름변경
- User + Service 혹은 Controller 및 dto 이름변경
