+++ BE/src/main/java/kr/codesquad/issuetracker/service/MemberService.java
@@ -6,7 +6,7 @@ import kr.codesquad.issuetracker.auth.dto.UserProfile;
 import kr.codesquad.issuetracker.auth.service.JwtService;
 import kr.codesquad.issuetracker.domain.member.Member;
 import kr.codesquad.issuetracker.domain.member.MemberRepository;
-import kr.codesquad.issuetracker.web.dto.user.UserResponseDto;
+import kr.codesquad.issuetracker.web.dto.member.MemberResponseDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -15,19 +15,19 @@ import org.springframework.transaction.annotation.Transactional;
 @Slf4j
 @RequiredArgsConstructor
 @Service
-public class UserService {
+public class MemberService {
 
 	private final GithubOauth oauth;
 	private final JwtService jwtService;
 	private final MemberRepository memberRepository;
 
 	@Transactional
-	public UserResponseDto login(String code) {
+	public MemberResponseDto login(String code) {
 		AccessTokenResponseDto token = oauth.getToken(code);
 		log.debug("tokenResponse = {}", token);
 		UserProfile userInfo = oauth.getUserInfo(token.getAccessToken());
 		log.debug("userInfo = {}", userInfo);
 		Member member = Member.createMember(userInfo, token.getAccessToken());
-		return UserResponseDto.of(member, jwtService.createToken(memberRepository.save(member)));
+		return MemberResponseDto.of(member, jwtService.createToken(memberRepository.save(member)));
 	}
 }
