+++ BE/src/main/java/kr/codesquad/issuetracker/web/dto/member/MemberResponseDto.java
@@ -1,23 +1,22 @@
-package kr.codesquad.issuetracker.web.dto.user;
+package kr.codesquad.issuetracker.web.dto.member;
 
 import kr.codesquad.issuetracker.domain.member.Member;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.ToString;
 
 @ToString
 @Getter
 @AllArgsConstructor
-public class UserResponseDto {
+public class MemberResponseDto {
 
 	private final String name;
 	private final String email;
 	private final String avatarUrl;
 	private final String token;
 
-	public static UserResponseDto of(Member member, String token) {
-		return new UserResponseDto(
+	public static MemberResponseDto of(Member member, String token) {
+		return new MemberResponseDto(
 			member.getName(),
 			member.getEmail(),
 			member.getImageUrl(),
